Korrekturen + einfaches Delete hinzugefügt

+++ backend onlineshop/src/main/java/de/vkb/onlineshop/service/KundeService.java
@@ -29,8 +29,12 @@ public class KundeService {
         return repository.save(kunde);
     }
 
-    public void delete(int id) {
+    public boolean deleteById(int id) {
+        if (!repository.existsById(id)) {
+            return false;
+        }
         repository.deleteById(id);
+        return true;
     }
 
 }
